Use process module for console commands
Replace the old system to run commands. Instead of using bash or cmd
on windows and to use temporary files for input/output we use now
the process module which is based on the reproc library.

Once this change is consolidated the functions to create temporary files
may be removed from the core if no other plugin use them.