Plug console view only when required
To avoid a race between plugins we insert the console view in the
nodes' hierarchy only when the console is actually needed. In this
way we ensure that the console view is always embedded below the
documents' view.

Otherwise the console view may appear below the treeview if the console
plugin is loaded before the treeview.